Add uring_poll_fn() API for timeout-based polling
- Add new uring_poll_fn() function in src/uring_async.rs:
  - Takes io_uring ring and timeout_secs parameter
  - Uses submit_with_args() with timespec for timeout detection
  - Returns bool indicating if timeout occurred (ETIME error)
  - Generic over EntryMarker for ring compatibility

- Update examples/null.rs to use new API:
  - Import uring_poll_fn in use statement
  - Replace inline timeout logic with uring_poll_fn(r, 20) call
  - Simplifies timeout handling and makes it reusable
